+++ api/context.js
@@ -1,3 +1,0 @@
-export default async function(req, context) {
-	return Response.json(context);
-}